[REFACTOR] RacingGame과 RacingController 합치기
